Add price and studio fields to Item and DTO models
Introduced 'price' and 'studio' fields to Item, Game, and ItemDTO classes to support type-specific attributes for GAME and CARTOON items. Updated related getters/setters and constructors. Minor repository and template formatting changes included.

+++ src/main/java/com/bookweb/model/Game.java
@@ -9,23 +9,24 @@ import jakarta.persistence.Entity;
 @DiscriminatorValue("GAME")
 public class Game extends Item {
 
-	private int price;
+	private Double price;
 
 	public Game() {
 		super();
 		this.setType(ItemType.GAME);
 	}
 
-	public Game(ItemType type, String title, String creator, String description, Set<Category> categories, int price) {
+	public Game(ItemType type, String title, String creator, String description, Set<Category> categories,
+			Double price) {
 		super(ItemType.GAME, title, creator, description, categories);
 		this.price = price;
 	}
 
-	public int getPrice() {
+	public Double getPrice() {
 		return price;
 	}
 
-	public void setPrice(int price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
 }
